src/pyDANETLSA/requirements.txt
cryptography
pyOpenSSL
requests
dnspython